rename announcer_chair to cli_display
